requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
psycopg2-binary==2.9.7
cryptography==41.0.3
pandas==2.0.3
plotly==5.16.1
scikit-learn==1.3.0
numpy==1.24.3
lightgbm==4.0.0
werkzeug==2.3.7